Modificacion clase MainWindow (Correcion de acceso a cargas de las barras) - Modificaion clase Jugador (Implementacion de algunos gets y sets) - Implementacion inicial de clase Enemigo

+++ proyectoFinal/personaje.cpp
@@ -29,7 +29,7 @@ Personaje::Personaje(QPixmap _hojaSprite,
     setPixmap(spriteEscalado);
 }
 
-int Personaje::getCargaVida() const {return cargaVida;}
+unsigned short Personaje::getCargaVida() const {return cargaVida;}
 
 void Personaje::movimientoSprite(int direccion, unsigned short cantidadDeSprites)
 {
